Fix exactOptionalPropertyTypes, fix TS 5.4
https://github.com/sindresorhus/ky/pull/559 undid the fix in
https://github.com/sindresorhus/ky/pull/543. Added
`exactOptionalPropertyTypes` to `compilerOptions` to prevent this from
happening again, seems pretty low-impact. It should be okay to pass
`undefined` to `credentials` and other options, but this is an upstream
problem (see `RequestInit` in `undici-types/fetch.d.ts`).

The build was failing on TS 5.4 because there's a `priority` in
`RequestInit` now.